backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
